+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release Notes
 
-## v0.0.7
-- In progress
+## v0.1.0
+- Netmouth SIP TTS service started
+- Start of libeXosip2 under golang support
 
 ## v0.0.6
 - Standardized gnu paths